fix format '%lld' warning in x86_64 platform
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/branches/2.3.2@4748 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ server/gfmd/inode.c
@@ -2299,22 +2299,25 @@ nlink_check(void *closure, struct inode *inode)
 			gflog_warning(GFARM_MSG_UNFIXED,
 			    "directory inode %lld nlink %lld should be 2 "
 			    "(really %lld): fixed",
-			    inode_get_number(inode), inode_get_nlink(inode),
-			    inode_get_nlink_ini(inode));
+			    (unsigned long long)inode_get_number(inode),
+			    (unsigned long long)inode_get_nlink(inode),
+			    (unsigned long long)inode_get_nlink_ini(inode));
 			inode->i_nlink = 2;
 		}
 		else
 			return;
 	} else if (inode_get_nlink_ini(inode) == 0) {
 		gflog_warning(GFARM_MSG_UNFIXED,
-		    "inode %lld is not referenced", inode_get_number(inode));
+		    "inode %lld is not referenced",
+		    (unsigned long long)inode_get_number(inode));
 		/* XXX - move to the /lost+found directory */
 		return;
 	} else if (inode_get_nlink(inode) != inode_get_nlink_ini(inode)) {
 		gflog_warning(GFARM_MSG_UNFIXED,
 		    "inode %lld nlink %lld should be %lld: fixed",
-		    inode_get_number(inode), inode_get_nlink(inode),
-		    inode_get_nlink_ini(inode));
+		    (unsigned long long)inode_get_number(inode),
+		    (unsigned long long)inode_get_nlink(inode),
+		    (unsigned long long)inode_get_nlink_ini(inode));
 		inode->i_nlink = inode_get_nlink_ini(inode);
 	}
 	e = db_inode_nlink_modify(
